updates for run optimization
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@7187 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 57c02dbcfc5e6b6bf9cf25b51496b1e55b354f53

+++ src/prerna/ui/components/specific/tap/SysSiteOptPlaySheet.java
@@ -117,8 +117,8 @@ public class SysSiteOptPlaySheet extends BasicProcessingPlaySheet{
 		String optType = gson.fromJson(gson.toJson(webDataHash.get("optTypes")), String.class);
 		
 		//advanced params
-		int infl = gson.fromJson(gson.toJson(webDataHash.get("infl")), Integer.class);
-		int disc = gson.fromJson(gson.toJson(webDataHash.get("disc")), Integer.class);
+		double infl = gson.fromJson(gson.toJson(webDataHash.get("infl")), Double.class);
+		double disc = gson.fromJson(gson.toJson(webDataHash.get("disc")), Double.class);
 		int numPts = gson.fromJson(gson.toJson(webDataHash.get("numPts")), Integer.class);
 		
 		opt = new SysSiteOptimizer();
